Move 'inChoiceListShown' from BaseTemplateBean to Process
