+++ Kitodo-DataManagement/src/main/resources/db/migration/V2_106__Remove_inChoiceListShown_column_from_templates.sql
@@ -1,0 +1,13 @@
+--
+-- (c) Kitodo. Key to digital objects e. V. <contact@kitodo.org>
+--
+-- This file is part of the Kitodo project.
+--
+-- It is licensed under GNU General Public License version 3 or later.
+--
+-- For the full copyright and license information, please read the
+-- GPL3-License.txt file that was distributed with this source code.
+--
+
+-- Remove inChoiceListShown column from template table
+ALTER TABLE template DROP COLUMN inChoiceListShown;
