+++ Kitodo/src/main/webapp/WEB-INF/templates/includes/processFromTemplate/dialogs/selectTemplate.xhtml
@@ -37,7 +37,7 @@
                         <f:selectItem itemValue="#{null}"
                                       itemLabel="#{msgs['selectTemplate']}"
                                       noSelectionOption="true"/>
-                        <f:selectItems value="#{SelectTemplateDialogView.project.availableTemplates}"
+                        <f:selectItems value="#{SelectTemplateDialogView.project.templates}"
                                        var="template"
                                        itemLabel="#{template.title}"
                                        itemValue="#{template.id}"/>
