+++ Kitodo/src/main/webapp/WEB-INF/templates/includes/projects/projectList.xhtml
@@ -170,7 +170,7 @@
                                          columns="2"
                                          columnClasses="label, actionsColumn" >
                                 <p:repeat id="templates"
-                                          value="#{item.availableTemplates}"
+                                          value="#{item.templates}"
                                           var="template">
                                     <div class="project-template-action">
                                         <h:outputText value="#{template.title}"
